+++ config/widgets.php
@@ -10,7 +10,7 @@ return ['Widgets' => [
 		'MeCms.Posts::categories',
 		'MeCms.Posts::latest' => ['limit' => 10],
 		'MeCms.Posts::search',
-		'MeCms.PostsTags::popular' => ['limit' => 10, 'shuffle' => TRUE, 'style' => ['maxFont' => 40, 'minFont' => 12]],
+		'MeCms.PostsTags::popular' => ['limit' => 10, 'prefix' => '#', 'shuffle' => TRUE, 'style' => ['maxFont' => 40, 'minFont' => 12]],
 	],
 	//Specific widgets for the homepage. 
 	//If empty, will be used the default widget