+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### 2.2.2
 * fixed bug in "album" and "posts categories" widgets;
 * widgets now use a common view. Rewritten the code of all widgets;
-* you can specify the minimum and maximum fonts for the "popular tags" widget.
+* you can specify the minimum font, the maximum font and the tag prefix for the "popular tags" widget.
 
 ### 2.2.1
 * added index for tags;